+++ crates/ryo3-tokio/src/lib.rs
@@ -1,6 +1,5 @@
 #![doc = include_str!("../README.md")]
 
-pub mod aiopen;
 pub mod fs;
 pub mod time;
 
@@ -15,6 +14,5 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     // fs
     fs::pymod_add(m)?;
 
-    aiopen::pymod_add(m)?;
     Ok(())
 }
