+++ python/ry/__init__.py
@@ -67,15 +67,18 @@ from ry.ryo3 import (
     bzip2_encode,
     camel_case,
     canonicalize,
+    canonicalize_async,
     cd,
     copy,
     copy_async,
     create_dir,
     create_dir_all,
+    create_dir_all_async,
     create_dir_async,
     date,
     datetime,
     dirs,
+    exists_async,
     fetch,
     fmt_size,
     fnv1a,
@@ -85,6 +88,7 @@ from ry.ryo3 import (
     gzip,
     gzip_decode,
     gzip_encode,
+    hard_link_async,
     home,
     instant,
     is_gzipped,
@@ -110,10 +114,13 @@ from ry.ryo3 import (
     read_dir,
     read_dir_async,
     read_json,
+    read_link_async,
     read_stream,
     read_text,
+    read_to_string_async,
     remove_dir,
     remove_dir_all,
+    remove_dir_all_async,
     remove_dir_async,
     remove_file,
     remove_file_async,
@@ -132,6 +139,7 @@ from ry.ryo3 import (
     timespan,
     title_case,
     train_case,
+    try_exists_async,
     unindent,
     unindent_bytes,
     uuid,
@@ -214,15 +222,18 @@ __all__ = (
     "bzip2_encode",
     "camel_case",
     "canonicalize",
+    "canonicalize_async",
     "cd",
     "copy",
     "copy_async",
     "create_dir",
     "create_dir_all",
+    "create_dir_all_async",
     "create_dir_async",
     "date",
     "datetime",
     "dirs",
+    "exists_async",
     "fetch",
     "fmt_size",
     "fnv1a",
@@ -232,6 +243,7 @@ __all__ = (
     "gzip",
     "gzip_decode",
     "gzip_encode",
+    "hard_link_async",
     "home",
     "instant",
     "is_gzipped",
@@ -257,10 +269,13 @@ __all__ = (
     "read_dir",
     "read_dir_async",
     "read_json",
+    "read_link_async",
     "read_stream",
     "read_text",
+    "read_to_string_async",
     "remove_dir",
     "remove_dir_all",
+    "remove_dir_all_async",
     "remove_dir_async",
     "remove_file",
     "remove_file_async",
@@ -280,6 +295,7 @@ __all__ = (
     "timespan",
     "title_case",
     "train_case",
+    "try_exists_async",
     "unindent",
     "unindent_bytes",
     "uuid",
