+++ python/ry/ryo3/__init__.pyi
@@ -128,15 +128,23 @@ from ._std import write_text as write_text
 from ._tokio import AsyncFile as AsyncFile
 from ._tokio import aiopen as aiopen
 from ._tokio import asleep as asleep
+from ._tokio import canonicalize_async as canonicalize_async
 from ._tokio import copy_async as copy_async
+from ._tokio import create_dir_all_async as create_dir_all_async
 from ._tokio import create_dir_async as create_dir_async
+from ._tokio import exists_async as exists_async
+from ._tokio import hard_link_async as hard_link_async
 from ._tokio import metadata_async as metadata_async
 from ._tokio import read_async as read_async
 from ._tokio import read_dir_async as read_dir_async
+from ._tokio import read_link_async as read_link_async
+from ._tokio import read_to_string_async as read_to_string_async
+from ._tokio import remove_dir_all_async as remove_dir_all_async
 from ._tokio import remove_dir_async as remove_dir_async
 from ._tokio import remove_file_async as remove_file_async
 from ._tokio import rename_async as rename_async
 from ._tokio import sleep_async as sleep_async
+from ._tokio import try_exists_async as try_exists_async
 from ._tokio import write_async as write_async
 from ._unindent import unindent as unindent
 from ._unindent import unindent_bytes as unindent_bytes
