+++ python/ry/ryo3/_tokio.pyi
@@ -14,14 +14,22 @@ from ry.ryo3._std import FileType, Metadata
 # =============================================================================
 # FS
 # =============================================================================
+async def canonicalize_async(path: FsPathLike) -> FsPathLike: ...
 async def copy_async(src: FsPathLike, dst: FsPathLike) -> None: ...
 async def create_dir_async(path: FsPathLike) -> None: ...
+async def create_dir_all_async(path: FsPathLike) -> None: ...
+async def hard_link_async(src: FsPathLike, dst: FsPathLike) -> None: ...
 async def metadata_async(path: FsPathLike) -> None: ...
 async def read_async(path: FsPathLike) -> Bytes: ...
 async def remove_dir_async(path: FsPathLike) -> None: ...
+async def remove_dir_all_async(path: FsPathLike) -> None: ...
 async def remove_file_async(path: FsPathLike) -> None: ...
+async def read_link_async(path: FsPathLike) -> FsPathLike: ...
+async def read_to_string_async(path: FsPathLike) -> str: ...
 async def rename_async(src: FsPathLike, dst: FsPathLike) -> None: ...
 async def write_async(path: FsPathLike, data: Buffer) -> None: ...
+async def try_exists_async(path: FsPathLike) -> bool: ...
+async def exists_async(path: FsPathLike) -> bool: ...
 
 class DirEntryAsync:
     def __fspath__(self) -> str: ...
